update android project and native lib location

+++ Android_Build/OpenPGP.Android.Wrapper/jni/include/utilities/BBS.h
@@ -56,7 +56,7 @@ public:
     std::string rand_byts(const unsigned int & bits = 1, const std::string & par = "even");
     std::string rand_b(const unsigned int & bits);
     std::string randbit();
-    void c_rand(unsigned char *dest, size_t length);
+    void c_rand(unsigned char *dest, int length);
     
     std::string gen_random_key(const int len);
 };
